fix: move GQL into a single file
